+++ Tests/AppTests/Disabled/ArchitecturePerformanceComparisonTests.swift
