+++ Tests/AppTests/Disabled/ServiceRegistryPerformanceTests.swift
